# web-client/server/docs/swagger.json
{
	"basePath": "/",
	"definitions": {
...
		"Vendor": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"files": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/VendorFiles"
					},
					"description": ""
				},
				"known": {
					"type": "boolean",
					"description": ""
				},
				"group_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"circuits": {
					"type": "object",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"files",
				"known",
				"group_id",
				"circuits"
			]
		},
		"ValueScoreMetrics": {
			"description": "",
			"properties": {
				"market": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"packetLoss": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"latency": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"uptime": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"utilizationSend": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"utilizationReceive": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"market",
				"packetLoss",
				"latency",
				"uptime",
				"utilizationSend",
				"utilizationReceive"
			]
		},
...
		"SlaViolations": {
			"description": "",
			"properties": {
				"availability": {
					"$ref": "#/definitions/SlaCheck"
				},
				"throughputSend": {
					"$ref": "#/definitions/SlaCheck"
				},
				"throughputReceive": {
					"$ref": "#/definitions/SlaCheck"
				},
				"jitter": {
					"$ref": "#/definitions/SlaCheck"
				},
				"packetLoss": {
					"$ref": "#/definitions/SlaCheck"
				},
				"latency": {
					"$ref": "#/definitions/SlaCheck"
				}
			},
			"type": "object",
			"required": [
				"availability",
				"throughputSend",
				"throughputReceive",
				"jitter",
				"packetLoss",
				"latency"
			]
		},
		"Circuit": {
			"description": "",
			"properties": {
				"circuit_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"host_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"item_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"owner_account_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"sla_availability": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"sla_throughput_send": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"sla_throughput_receive": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"sla_jitter": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"sla_packet_loss": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"sla_latency": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"cost": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"term": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"remote_ip": {
					"type": "string",
					"description": ""
				},
				"remote_host_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"vendor": {
					"$ref": "#/definitions/Vendor"
				},
				"creationDate": {
					"type": "string",
					"format": "date-time",
					"description": ""
				},
				"expectedMonthlyCost": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"monitoringData": {
					"type": "object",
					"description": ""
				},
				"_valueMetrics": {
					"$ref": "#/definitions/ValueScoreMetrics"
				},
				"_valueScore": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"_slaViolations": {
					"$ref": "#/definitions/SlaViolations"
				}
			},
			"type": "object",
			"required": [
				"circuit_id",
				"host_id",
				"item_id",
				"owner_account_id",
				"name",
				"sla_availability",
				"sla_throughput_send",
				"sla_throughput_receive",
				"sla_jitter",
				"sla_packet_loss",
				"sla_latency",
				"cost",
				"term",
				"remote_ip",
				"remote_host_id",
				"vendor",
				"creationDate",
				"expectedMonthlyCost",
				"monitoringData",
				"_valueMetrics",
				"_valueScore",
				"_slaViolations"
			]
		},
...
		"DeviceInterface": {
			"description": "",
			"properties": {
				"interface_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"host_id": {
					"type": "string",
					"description": ""
				},
				"item_id": {
					"type": "string",
					"description": ""
				},
				"displayName": {
					"type": "string",
					"description": ""
				},
				"circuit_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"vendor_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"circuit": {
					"$ref": "#/definitions/Circuit"
				}
			},
			"type": "object",
			"required": [
				"interface_id",
				"host_id",
				"item_id",
				"displayName",
				"circuit_id",
				"vendor_id",
				"circuit"
			]
		},
		"Address": {
			"description": "",
			"properties": {
				"LineOne": {
					"type": "string",
					"description": ""
				},
				"LineTwo": {
					"type": "string",
					"description": ""
				},
				"City": {
					"type": "string",
					"description": ""
				},
				"State": {
					"type": "string",
					"description": ""
				},
				"Zip": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"LineOne",
				"LineTwo",
				"City",
				"State",
				"Zip"
			]
		},
		"Geolocation": {
			"description": "",
			"properties": {},
			"type": "object"
		},
		"History": {
			"description": "",
			"properties": {
				"length": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"state": {
					"type": "object",
					"description": ""
				},
				"scrollRestoration": {
					"type": "string",
					"enum": [
						"auto",
						"manual"
					],
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"length",
				"state",
				"scrollRestoration"
			]
		},
		"Item": {
			"description": "A vector of data collected by network monitoring\nBelongs to a Device or Interface\n\nhttps://www.zabbix.com/documentation/3.4/manual/api/reference/item",
			"properties": {
				"key": {
					"type": "string",
					"description": "Item key\nZabbix - [[Item.key_]]"
				},
				"name": {
					"type": "string",
					"description": "Name of the Metric\nZabbix - [[Item.Name]]"
				},
				"metricType": {
					"type": "string",
					"enum": [
						"Zabbix agent",
						"SNMPv1 agent",
						"Zabbix trapper",
						"simple check",
						"SNMPv2 agent",
						"Zabbix internal",
						"Zabbix internal",
						"Zabbix internal",
						"SNMPv3 agent",
						"Zabbix agent (active)",
						"Zabbix aggregate",
						"web item",
						"external check",
						"database monitor",
						"IPMI agent",
						"SSH agent",
						"TELNET agent",
						"calculated",
						"JMX agent",
						"SNMP agent",
						"Dependent item"
					],
					"description": "Type of Metric / Metric collection source\nZabbix - [[Item.type]]"
				},
				"valueType": {
					"type": "string",
					"enum": [
						"numeric float",
						"character",
						"log",
						"numberic unsigned",
						"text"
					],
					"description": "Type of value the metric provides\nZabbix - [[Item.value_type]]"
				},
				"description": {
					"type": "string",
					"description": "Description of the metric\nZabbix - [[Item.description]]"
				},
				"error": {
					"type": "string",
					"description": "Error text if there are problems updating the item\nZabbix - [[Item.error]]"
				},
				"history": {
					"type": "string",
					"description": "A time unit of how long the history should be stored. Also accepts user macro\nZabbix - [[Item.history]]"
				},
				"lastValue": {
					"type": "object",
					"description": "Last value of the metric\nZabbix - [[Item.lastvalue]]"
				},
				"isSupported": {
					"type": "boolean",
					"description": "Whether or not the metric is supported by Zabbix\nZabbix - [[Item.State]]"
				},
				"units": {
					"type": "string",
					"description": "Units the Metric value is measured in\nZabbix - [[Item.units]]"
				},
				"itemid": {
					"type": "string",
					"description": "Zabbix - [[Item.itemid]]"
				},
				"delay": {
					"type": "string",
					"description": "Updatnterval of the item. Accepts seconds or time\nunit with suffix and with or without one or more custom\nintervals that consist of either flexible intervals and\nscheduling intervals as serialized strings. Also accepts\nuser macros. Flexible intervals could be written as two\nmacros separated by a forward slash. Intervals are\nseparated by a semicolon.\nOptional for Zabbix trapper or Dependent iteme interval\nof the item. Accepts seconds or time unit with suffix\nand with or without one or more custom intervals that\nconsist of either flexible intervals and scheduling\nintervals as serialized strings. Also accepts user macros.\nFlexible intervals could be written as two macros separated\nby a forward slash. Intervals are separated by a semicolon.\nZabbix - [[Item.delay]]"
				},
				"hostId": {
					"type": "string",
					"description": "[[Device.zabbixHostId]]\nZabbix - [[Host.hostid]]"
				},
				"interfaceId": {
					"type": "string",
					"description": "[[Interface.interfaceId]]\nZabbix - [[Interface.interfaceid]]"
				},
				"type": {
					"type": "string",
					"description": ""
				},
				"value_type": {
					"type": "string",
					"description": ""
				},
				"historys": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/History"
					},
					"description": "Metric data obtained through the Zabbix [[History]] API"
				}
			},
			"type": "object",
			"required": [
				"key",
				"name",
				"metricType",
				"valueType",
				"description",
				"error",
				"history",
				"lastValue",
				"isSupported",
				"units",
				"itemid",
				"delay",
				"hostId",
				"interfaceId",
				"type",
				"value_type",
				"historys"
			]
		},
		"Graph": {
			"description": "",
			"properties": {
				"groupIds": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"graphId": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"width": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"height": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"yAxisMinimum": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"yAxisMaximum": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"items": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Item"
					},
					"description": ""
				},
				"gitems": {
					"type": "array",
					"items": {
						"type": "object"
					},
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"groupIds",
				"graphId",
				"name",
				"width",
				"height",
				"yAxisMinimum",
				"yAxisMaximum",
				"items",
				"gitems"
			]
		},
		"Metric": {
			"description": "",
			"properties": {
				"key": {
					"type": "string",
					"description": "Item key\nZabbix - [[Item.key_]]"
				},
				"name": {
					"type": "string",
					"description": "Name of the Metric\nZabbix - [[Item.Name]]"
				},
				"metricType": {
					"type": "string",
					"enum": [
						"Zabbix agent",
						"SNMPv1 agent",
						"Zabbix trapper",
						"simple check",
						"SNMPv2 agent",
						"Zabbix internal",
						"Zabbix internal",
						"Zabbix internal",
						"SNMPv3 agent",
						"Zabbix agent (active)",
						"Zabbix aggregate",
						"web item",
						"external check",
						"database monitor",
						"IPMI agent",
						"SSH agent",
						"TELNET agent",
						"calculated",
						"JMX agent",
						"SNMP agent",
						"Dependent item"
					],
					"description": "Type of Metric / Metric collection source\nZabbix - [[Item.type]]"
				},
				"valueType": {
					"type": "string",
					"enum": [
						"numeric float",
						"character",
						"log",
						"numberic unsigned",
						"text"
					],
					"description": "Type of value the metric provides\nZabbix - [[Item.value_type]]"
				},
				"description": {
					"type": "string",
					"description": "Description of the metric\nZabbix - [[Item.description]]"
				},
				"error": {
					"type": "string",
					"description": "Error text if there are problems updating the item\nZabbix - [[Item.error]]"
				},
				"history": {
					"type": "string",
					"description": "A time unit of how long the history should be stored. Also accepts user macro\nZabbix - [[Item.history]]"
				},
				"lastValue": {
					"type": "object",
					"description": "Last value of the metric\nZabbix - [[Item.lastvalue]]"
				},
				"isSupported": {
					"type": "boolean",
					"description": "Whether or not the metric is supported by Zabbix\nZabbix - [[Item.State]]"
				},
				"units": {
					"type": "string",
					"description": "Units the Metric value is measured in\nZabbix - [[Item.units]]"
				},
				"itemid": {
					"type": "string",
					"description": "Zabbix - [[Item.itemid]]"
				},
				"delay": {
					"type": "string",
					"description": "Updatnterval of the item. Accepts seconds or time\nunit with suffix and with or without one or more custom\nintervals that consist of either flexible intervals and\nscheduling intervals as serialized strings. Also accepts\nuser macros. Flexible intervals could be written as two\nmacros separated by a forward slash. Intervals are\nseparated by a semicolon.\nOptional for Zabbix trapper or Dependent iteme interval\nof the item. Accepts seconds or time unit with suffix\nand with or without one or more custom intervals that\nconsist of either flexible intervals and scheduling\nintervals as serialized strings. Also accepts user macros.\nFlexible intervals could be written as two macros separated\nby a forward slash. Intervals are separated by a semicolon.\nZabbix - [[Item.delay]]"
				},
				"hostId": {
					"type": "string",
					"description": "[[Device.zabbixHostId]]\nZabbix - [[Host.hostid]]"
				},
				"interfaceId": {
					"type": "string",
					"description": "[[Interface.interfaceId]]\nZabbix - [[Interface.interfaceid]]"
				},
				"type": {
					"type": "string",
					"description": ""
				},
				"value_type": {
					"type": "string",
					"description": ""
				},
				"historys": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/History"
					},
					"description": "Metric data obtained through the Zabbix [[History]] API"
				}
			},
			"type": "object",
			"required": [
				"key",
				"name",
				"metricType",
				"valueType",
				"description",
				"error",
				"history",
				"lastValue",
				"isSupported",
				"units",
				"itemid",
				"delay",
				"hostId",
				"interfaceId",
				"type",
				"value_type",
				"historys"
			]
		},
		"Notification": {
			"description": "",
			"properties": {
				"body": {
					"type": "string",
					"description": ""
				},
				"dir": {
					"type": "string",
					"enum": [
						"auto",
						"ltr",
						"rtl"
					],
					"description": ""
				},
				"icon": {
					"type": "string",
					"description": ""
				},
				"lang": {
					"type": "string",
					"description": ""
				},
				"permission": {
					"type": "string",
					"enum": [
						"default",
						"denied",
						"granted"
					],
					"description": ""
				},
				"tag": {
					"type": "string",
					"description": ""
				},
				"title": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"body",
				"dir",
				"icon",
				"lang",
				"permission",
				"tag",
				"title"
			]
		},
		"HostInterface": {
			"description": "",
			"properties": {
				"hostInterfaceId": {
					"type": "string",
					"description": ""
				},
				"dnsName": {
					"type": "string",
					"description": ""
				},
				"hostId": {
					"type": "string",
					"description": ""
				},
				"ipAddress": {
					"type": "string",
					"description": ""
				},
				"isPrimaryInterface": {
					"type": "boolean",
					"description": ""
				},
				"port": {
					"type": "string",
					"description": ""
				},
				"type": {
					"type": "string",
					"enum": [
						"1",
						"2",
						"3",
						"4"
					],
					"description": ""
				},
				"shouldConnectViaIp": {
					"type": "boolean",
					"description": ""
				},
				"shouldUseBulkSnmpRequests": {
					"type": "boolean",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"hostInterfaceId",
				"dnsName",
				"hostId",
				"ipAddress",
				"isPrimaryInterface",
				"port",
				"type",
				"shouldConnectViaIp",
				"shouldUseBulkSnmpRequests"
			]
		},
		"HostItem": {
			"description": "",
			"properties": {
				"itemid": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"monitoringData": {
					"type": "object",
					"description": ""
				},
				"InterfaceMetadata": {
					"$ref": "#/definitions/DeviceInterface"
				},
				"Circuit": {
					"$ref": "#/definitions/Circuit"
				},
				"IsCircuit": {
					"type": "boolean",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"itemid",
				"name",
				"monitoringData",
				"InterfaceMetadata",
				"Circuit",
				"IsCircuit"
			]
		},
		"Iface": {
			"description": "",
			"properties": {
				"ownerAccountId": {
					"type": "string",
					"description": ""
				},
				"hostId": {
					"type": "string",
					"description": ""
				},
				"itemId": {
					"type": "string",
					"description": ""
				},
				"displayName": {
					"type": "string",
					"description": ""
				},
				"circuits": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Circuit"
					},
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"ownerAccountId",
				"hostId",
				"itemId",
				"displayName",
				"circuits"
			]
		},
		"ManagedDevice": {
			"description": "",
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"host_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"group_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"isManaged": {
					"type": "boolean",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"host_id",
				"group_id",
				"isManaged"
			]
		},
		"Host": {
			"description": "A network host device\n\nhttps://www.zabbix.com/documentation/3.4/manual/api/reference/host",
			"properties": {
				"zabbixHostId": {
					"type": "string",
					"description": "Zabbix - [[Host.hostid]]"
				},
				"groupIds": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": "Zabbix - [[ Host.groups[].groupid ]]"
				},
				"internalName": {
					"type": "string",
					"description": "Technical Name - set by the NOC or the device itself\nZabbix - [[Host.host]]"
				},
				"displayName": {
					"type": "string",
					"description": "Visible Name - set by the application user\nZabbix - [[Host.name]]"
				},
				"zabbixStatus": {
					"type": "string",
					"description": "Zabbix Monitoring Status\nZabbix - [[Host.status]]"
				},
				"deviceType": {
					"type": "string",
					"description": "Operating System\nZabbix - [[HostInventory.os]]"
				},
				"operatingSystem": {
					"type": "string",
					"description": "Operating System\nZabbix - [[HostInventory.os]]"
				},
				"serialNumberA": {
					"type": "string",
					"description": "Primary Serial Number of the device\nZabbix - [[HostInventory.serialno_a]]"
				},
				"serialNumberB": {
					"type": "string",
					"description": "Secondary Serial Number of the device\nZabbix - [[HostInventory.serialno_b]]"
				},
				"hardware": {
					"type": "string",
					"description": "Physical hardware the device is running\nZabbix - [[HostInventory.hardware]]"
				},
				"contact": {
					"type": "string",
					"description": "Point of contact for the device\nZabbix - [[HostInventory.contact]]"
				},
				"model": {
					"type": "string",
					"description": "Model number or Model Name of the device\nZabbix - [[HostInventory.model]]"
				},
				"siteAddressA": {
					"$ref": "#/definitions/Address"
				},
				"siteAddressB": {
					"$ref": "#/definitions/Address"
				},
				"siteAddressC": {
					"$ref": "#/definitions/Address"
				},
				"locationDescription": {
					"type": "string",
					"description": "Location description reported by the device\nZabbix - [[HostInventory.location]]"
				},
				"geolocation": {
					"$ref": "#/definitions/Geolocation"
				},
				"notes": {
					"type": "string",
					"description": "Zabbix - [[HostInventory.notes]]"
				},
				"graphs": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Graph"
					},
					"description": "Zabbix - [[Item]]"
				},
				"metrics": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Metric"
					},
					"description": "Zabbix - [[Item]]"
				},
				"ipAddress": {
					"type": "string",
					"description": "Zabbix - [[Interface.ip]]"
				},
				"service": {
					"type": "string",
					"description": ""
				},
				"triggers": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Notification"
					},
					"description": ""
				},
				"hostInterfaces": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/HostInterface"
					},
					"description": ""
				},
				"hostItems": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/HostItem"
					},
					"description": ""
				},
				"ifaces": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Iface"
					},
					"description": ""
				},
				"isBillable": {
					"type": "boolean",
					"description": "Whether to enable Network Assure billing for the Device"
				},
				"managedDevice": {
					"$ref": "#/definitions/ManagedDevice"
				}
			},
			"type": "object",
			"required": [
				"zabbixHostId",
				"groupIds",
				"internalName",
				"displayName",
				"zabbixStatus",
				"deviceType",
				"operatingSystem",
				"serialNumberA",
				"serialNumberB",
				"hardware",
				"contact",
				"model",
				"siteAddressA",
				"siteAddressB",
				"siteAddressC",
				"locationDescription",
				"geolocation",
				"notes",
				"graphs",
				"metrics",
				"ipAddress",
				"service",
				"triggers",
				"hostInterfaces",
				"hostItems",
				"ifaces",
				"isBillable",
				"managedDevice"
			]
		},
...
		"DeviceManagementHistory": {
			"description": "",
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"host_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"destinationManagedState": {
					"type": "boolean",
					"description": ""
				},
				"changeDate": {
					"type": "string",
					"format": "date-time",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"host_id",
				"destinationManagedState",
				"changeDate"
			]
		},
		"ManagedDeviceViewModel": {
			"description": "",
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"host_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"group_id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"isManaged": {
					"type": "boolean",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"host_id",
				"group_id",
				"isManaged"
			]
		},
...
	},
	"info": {
		"description": "This is the backend for Affiniti's Network Assure.",
		"license": {
			"name": "ISC"
		},
		"title": "Network Assure REST API",
		"version": "1.0"
	},
	"paths": {
...
		"/host/{hostId}": {
			"get": {
				"operationId": "HostGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Host"
							}
						}
					}
				},
				"description": "",
				"tags": [
					"Hosts"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "hostId",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		},
...
		"/managed-device": {
			"post": {
				"operationId": "ManagedDeviceCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/DeviceManagementHistory"
						}
					}
				},
				"description": "",
				"tags": [
					"Managed Devices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "viewModel",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ManagedDeviceViewModel"
						}
					}
				]
			},
			"get": {
				"operationId": "ManagedDeviceGetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"type": "object"
							}
						}
					}
				},
				"description": "",
				"tags": [
					"Managed Devices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
		"/managed-device/host/{device_id}": {
			"put": {
				"operationId": "ManagedDeviceUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"description": "",
				"tags": [
					"Managed Devices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "viewModel",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ManagedDeviceViewModel"
						}
					}
				]
			}
		},
		"/managed-device/host/{deviceId}": {
			"get": {
				"operationId": "ManagedDeviceFindOneByHostId",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ManagedDevice"
						}
					}
				},
				"description": "",
				"tags": [
					"Managed Devices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "deviceId",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		},
		"/managed-device/counts": {
			"get": {
				"operationId": "ManagedDeviceGetManagedDevicesCount",
				"produces": [
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "",
				"tags": [
					"Managed Devices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"swagger": "2.0",
	"securityDefinitions": {
		"bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "headers"
		}
	},
	"host": "localhost:5050"
}